chore: renamed folder + namespace to avoid conflict with the other login page
