+++ Xappy/Xappy/Content/Scenarios/Login2/Controls/GradientView.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Xamarin.Forms;
 
-namespace Xappy.Content.Scenarios.Login.Controls
+namespace Xappy.Content.Scenarios.Login2.Controls
 {
     public enum GradientDirection
     {
