+++ Xappy/Xappy/Content/Scenarios/Login2/TaskExtensions.cs
@@ -3,7 +3,7 @@ using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 
-namespace Xappy.Content.Scenarios.Login
+namespace Xappy.Content.Scenarios.Login2
 {
     public static class TasksExtensions
     {
