+++ Xappy/Xappy/Xappy.csproj
@@ -37,8 +37,8 @@
     <Folder Include="Content\Scenarios\ToDo\" />
     <Folder Include="Content\About\" />
     <Folder Include="Content\Common\" />
-    <Folder Include="Content\Scenarios\Login\" />
-    <Folder Include="Content\Scenarios\Login\Controls\" />
+    <Folder Include="Content\Scenarios\Login2\" />
+    <Folder Include="Content\Scenarios\Login2\Controls\" />
   </ItemGroup>
   <ItemGroup>
     <Compile Remove="ApiConstants.cs" />
